Textpad 2
Adds a open file button and shows what file your opening. If you need a simple text editor, this one is only 3kb and can save and open files. If you need to save 2 kilobytes for whatever reason and don`t need a open file button, V1 is just that.